Added firebase account creation, user file creation, and basic email validation to NewAccountActivity.

+++ app/src/main/java/com/example/daniel/conconnection/User.java
@@ -1,7 +1,5 @@
 package com.example.daniel.conconnection;
 
-import java.io.Serializable;
-
 /**
  * Created by Daniel on 4/7/2016.
  */
@@ -9,7 +7,7 @@ public interface User{
     String getEmail();
     String getPassword();
     String getName();
-    short getAge();
+    int getAge();
     String getBodyType();
     String getBiography();
     boolean getAutoLogin();
@@ -17,7 +15,7 @@ public interface User{
     void setEmail(String email);
     void setPassword(String password);
     void setName(String name);
-    void setAge(short age);
+    void setAge(int age);
     void setBodyType(String bodyType);
     void setBiography(String biography);
     void setAutoLogin(boolean autoLogin);
